+++ src/sc_handle.rs
@@ -1,21 +1,21 @@
-use winapi::um::winsvc;
+use windows_sys::Win32::{Security, System::Services};
 
-/// A handle holder that wraps a low level [`winsvc::SC_HANDLE`].
-pub(crate) struct ScHandle(winsvc::SC_HANDLE);
+/// A handle holder that wraps a low level [`Security::SC_HANDLE`].
+pub(crate) struct ScHandle(Security::SC_HANDLE);
 
 impl ScHandle {
-    pub(crate) unsafe fn new(handle: winsvc::SC_HANDLE) -> Self {
+    pub(crate) unsafe fn new(handle: Security::SC_HANDLE) -> Self {
         ScHandle(handle)
     }
 
-    /// Returns underlying [`winsvc::SC_HANDLE`].
-    pub(crate) fn raw_handle(&self) -> winsvc::SC_HANDLE {
+    /// Returns underlying [`Security::SC_HANDLE`].
+    pub(crate) fn raw_handle(&self) -> Security::SC_HANDLE {
         self.0
     }
 }
 
 impl Drop for ScHandle {
     fn drop(&mut self) {
-        unsafe { winsvc::CloseServiceHandle(self.0) };
+        unsafe { Services::CloseServiceHandle(self.0) };
     }
 }