+++ CHANGELOG.md
@@ -6,7 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- Breaking: Change `winapi` dependency to `windows-sys`. This is a breaking change since some of the low level Windows types are exposed in the public API of this library.
+- Breaking: Change `winapi` dependency to `windows-sys`. This is a breaking change since
+  some of the low level Windows types are exposed in the public API of this library.
 - Breaking: Update `widestring` dependency to 1.0 and remove it from the public API.